third_party/blink/web_tests/platform/generic/fast/parser/block-nesting-cap-expected.txt

Test that the HTML parser does not allow the nesting depth of elements to exceed 512.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS d512.parentNode === d510 is true
PASS d511.parentNode === d510 is true
PASS d512.previousSibling === d511 is true
PASS successfullyParsed is true

TEST COMPLETE
